ref(all): Separate one big file to many small files
